backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic[email]==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
mysql-connector-python==8.3.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
pytest==8.0.2
httpx==0.27.0
email-validator==2.1.0.post1